requirements.txt
colorama~=0.4.6
json_canonical
pynacl
jsonschema